Updated Derelict to the current SVN version. Fixed DMD 2.057 compatibility errors. Minor bugfixes.

+++ dependencies/derelict/DerelictOgg/derelict/ogg/vorbisfiletypes.d
@@ -29,6 +29,7 @@ module derelict.ogg.vorbisfiletypes;
 
 private
 {
+    import derelict.util.compat;
     import derelict.ogg.oggtypes;
     import derelict.ogg.vorbistypes;
     import derelict.ogg.vorbisenctypes;
@@ -41,7 +42,7 @@ extern (C)
         size_t function(void* ptr, size_t size, size_t nmemb, void* datasource ) read_func;
         int function(void* datasource, ogg_int64_t offset, int whence ) seek_func;
         int function(void* datasource ) close_func;
-        int function(void* datasource ) tell_func;
+        c_long function(void* datasource ) tell_func;
     }
 }
 
@@ -63,13 +64,13 @@ struct OggVorbis_File
     int              links;
     ogg_int64_t     *offsets;
     ogg_int64_t     *dataoffsets;
-    int             *serialnos;
+    c_long             *serialnos;
     ogg_int64_t     *pcmlengths;
     vorbis_info     *vi;
     vorbis_comment  *vc;
     ogg_int64_t      pcm_offset;
     int              ready_state;
-    int              current_serialno;
+    c_long              current_serialno;
     int              current_link;
     double           bittrack;
     double           samptrack;
